src/pipelines/diversified_dataset_extractor/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
beartype==0.17.2
    # via picsellia
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecation==2.1.0
    # via picsellia
filelock==3.13.4
    # via
    #   huggingface-hub
    #   torch
fsspec==2024.3.1
    # via
    #   huggingface-hub
    #   torch
ftfy==6.2.0
    # via open-clip-torch
huggingface-hub==0.22.2
    # via
    #   open-clip-torch
    #   timm
idna==3.7
    # via requests
imagehash==4.3.1
jinja2==3.1.3
    # via torch
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   imagehash
    #   pywavelets
    #   scipy
    #   torchvision
open-clip-torch==2.24.0
orjson==3.10.1
    # via picsellia
packaging==24.0
    # via
    #   deprecation
    #   huggingface-hub
picsellia==6.14.2
picsellia-annotations==0.6.0
    # via picsellia
picsellia-connexion-services==0.2.1
    # via picsellia
pillow==9.5.0
    # via
    #   imagehash
    #   picsellia
    #   torchvision
protobuf==5.26.1
    # via open-clip-torch
pydantic==1.10.15
    # via
    #   picsellia
    #   picsellia-annotations
pywavelets==1.6.0
    # via imagehash
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   picsellia
    #   timm
regex==2024.4.28
    # via open-clip-torch
requests==2.31.0
    # via
    #   huggingface-hub
    #   picsellia
    #   picsellia-connexion-services
safetensors==0.4.3
    # via timm
scipy==1.13.0
    # via imagehash
sentencepiece==0.2.0
    # via open-clip-torch
sympy==1.12
    # via torch
tabulate==0.9.0
tdqm==0.0.1
    # via picsellia
timm==0.9.16
    # via open-clip-torch
torch==2.3.0
    # via
    #   open-clip-torch
    #   timm
    #   torchvision
torchvision==0.18.0
    # via
    #   open-clip-torch
    #   timm
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   open-clip-torch
    #   tdqm
typing-extensions==4.11.0
    # via
    #   huggingface-hub
    #   pydantic
    #   torch
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via ftfy
xmltodict==0.13.0
    # via picsellia-annotations
